+++ src/main/java/ai/lucidtech/las/sdk/UpdateDocumentOptions.java
@@ -4,37 +4,35 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 
 
-public class CreateDocumentOptions extends Options {
-    private String consentId;
+public class UpdateDocumentOptions extends Options {
     private String datasetId;
     private JSONArray groundTruth;
     private Integer retentionInDays;
 
-    public CreateDocumentOptions setConsentId(String consentId) {
-        this.consentId = consentId;
-        return this;
-    }
-
-    public CreateDocumentOptions setDatasetId(String datasetId) {
+    public UpdateDocumentOptions setDatasetId(String datasetId) {
         this.datasetId = datasetId;
         return this;
     }
 
-    public CreateDocumentOptions setGroundTruth(JSONArray groundTruth) {
+    public UpdateDocumentOptions setGroundTruth(JSONArray groundTruth) {
         this.groundTruth = groundTruth;
         return this;
     }
 
-    public CreateDocumentOptions setRetentionInDays(Integer retentionInDays) {
+    public UpdateDocumentOptions setRetentionInDays(Integer retentionInDays) {
         this.retentionInDays = retentionInDays;
         return this;
     }
 
     public JSONObject addOptions(JSONObject body) {
-        this.addOption(body, "consentId", this.consentId);
         this.addOption(body, "datasetId", this.datasetId);
         this.addOption(body, "groundTruth", this.groundTruth);
         this.addOption(body, "retentionInDays", this.retentionInDays);
         return body;
+    }
+
+    public JSONObject toJson() {
+        JSONObject body = new JSONObject();
+        return this.addOptions(body);
     }
 }